refactor(engine): extract command dispatch into engine_drain_commands
